Updated new service and updated endpoints and imports to fit the updated backend
